AND-8562 Rename Config and move to local.config package
