--- core/datasource/src/main/java/com/tangem/datasource/config/models/ExpressModel.kt
@@ -1,12 +1,0 @@
-package com.tangem.datasource.config.models
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class ExpressModel(
-    @Json(name = "apiKey")
-    val apiKey: String,
-    @Json(name = "signVerifierPublicKey")
-    val signVerifierPublicKey: String,
-)
